Add minor corrections to readme and testing.md, change conditions to view login link to perms from is_staff
Minor proofreading edits. Changes the condition for logged-in users being able to see the links to see the staff pages from if user.is_staff to if perms.events.view_event, so users with permissions do not necessarily need to be granted staff status as well (staff, as in, can log into the Django admin panel).